requirements.txt
streamlit==1.31.0
sentence-transformers==2.3.1
torch==2.1.2
torchvision==0.16.2
Pillow==10.2.0
transformers==4.37.2